+++ src/mqtt/TemperatureFormat.ts
@@ -1,4 +1,5 @@
 import mqtt = require('async-mqtt');
+import jsonpath = require('jsonpath');
 
 import { RinnaiTouchPlatform } from '../platform';
 import { IMqttFormat } from './MqttService';
@@ -11,7 +12,11 @@ export class TemperatureFormat implements IMqttFormat {
     private readonly client: mqtt.AsyncMqttClient,
   ) {
     if (platform.settings.mqtt!.subscribeTemperature) {
-      this.zoneTopics = <Record<string, string>>platform.settings.mqtt!.subscribeTemperature; 
+      for(const zone of this.platform.service.AllZones) {
+        if (platform.settings.mqtt!.subscribeTemperature[zone] !== undefined) {
+          this.zoneTopics[zone] = platform.settings.mqtt!.subscribeTemperature[zone];
+        }
+      }
     }
   }
 
@@ -25,11 +30,39 @@ export class TemperatureFormat implements IMqttFormat {
     try {
       for(const zone in this.zoneTopics) {
         if (this.zoneTopics[zone] === topic) {
-          this.platform.service.setCurrentTemperatureOverride(parseFloat(payload), zone);
+          const temperture = this.extractTemperature(zone, payload);
+          this.platform.log.info(`MQTT: Extracted Temperature: ${temperture}`);
+          if (temperture !== undefined) {
+            this.platform.service.setCurrentTemperatureOverride(temperture, zone);
+          }
         }
       }
     } catch(error) {
       this.platform.log.error(error);
+    }
+  }
+
+  private extractTemperature(zone: string, payload: string): number | undefined {
+    this.platform.log.debug(this.constructor.name, 'extractTemperature', zone, payload);
+
+    try {
+      let temperature: number;
+
+      const path = this.platform.settings.mqtt?.subscribeTemperature[`jsonPath${zone}`];
+      if (path === undefined || path === '') {
+        temperature = Number(payload);
+        return isNaN(temperature) ? undefined : temperature;
+      }
+
+      const json = JSON.parse(payload);
+      const result = jsonpath.query(json, path);
+
+      if (Array.isArray(result) && result.length > 0) {
+        temperature = Number(result[0]);
+        return isNaN(temperature) ? undefined : temperature;
+      }
+    } catch {
+      return;
     }
   }